backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
cohere==4.47
qdrant-client==1.7.3
python-dotenv==1.0.0
pydantic==2.6.0
psycopg[binary]==3.1.18
httpx==0.26.0
bcrypt==4.1.2
pyjwt==2.8.0
openai==1.10.0
